revert(RX DMA Calypte): deactivate relaxed ordering
This is a revert of a commit 6f255e2c7f46fb40e1703c3e335f9b598bbe84f2

The relaxed ordering causes some packets in the upstream direction to
not be registered at all, e.g. to be dropped. The ordering of the
transactions needs to be retained and therefore, this Relaxed ordering
setting in the PCIe header should never be activated again.